Agrega a los test de Facturación la eliminación de facturas de compra, y el control de stock al eliminar una factura de venta.
