+++ src/test/java/sic/modelo/dto/ProductoDTO.java
@@ -1,5 +1,6 @@
 package sic.modelo.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,19 +14,19 @@ public class ProductoDTO {
     private Long id_Producto = 0L;
     private String codigo = "ABC123";
     private String descripcion = "Cinta adhesiva doble faz 3M";
-    private double cantidad = 10;
-    private double cantMinima = 2;    
-    private double ventaMinima = 0;
+    private BigDecimal cantidad = BigDecimal.TEN;
+    private BigDecimal cantMinima = new BigDecimal("2");    
+    private BigDecimal ventaMinima = BigDecimal.ZERO;
     private String nombreMedida = "Unidad";
-    private double precioCosto = 100;
-    private double ganancia_porcentaje = 50;
-    private double ganancia_neto = 50;
-    private double precioVentaPublico = 150;
-    private double iva_porcentaje = 21;
-    private double iva_neto = 31.5;
-    private double impuestoInterno_porcentaje = 0;
-    private double impuestoInterno_neto = 0;
-    private double precioLista = 181.5;    
+    private BigDecimal precioCosto = new BigDecimal("100");
+    private BigDecimal ganancia_porcentaje = new BigDecimal("50");
+    private BigDecimal ganancia_neto = new BigDecimal("50");
+    private BigDecimal precioVentaPublico = new BigDecimal("150");
+    private BigDecimal iva_porcentaje = new BigDecimal("21");
+    private BigDecimal iva_neto = new BigDecimal("31.5");
+    private BigDecimal impuestoInterno_porcentaje = BigDecimal.ZERO;
+    private BigDecimal impuestoInterno_neto = BigDecimal.ZERO;
+    private BigDecimal precioLista = new BigDecimal("181.5");    
     private String nombreRubro = "Ferreteria";
     private boolean ilimitado = false;
     private Date fechaUltimaModificacion = new Date(1463540400000L); // 18-05-2016
